move elastic query builder function for transaction filter
